// src/student/model/vo/StudentCountType.ts
export enum StudentCountType {
  NEW = 'NEW',
  PASSED = 'PASSED',
  APPROVAL = 'APPROVAL',
  WAITING = 'WAITING',
  NEW_WAITING = 'NEW_WAITING',
  EVERY = 'EVERY',
  LEARNING_STATE = 'LEARNING_STATE',
  PROPOSAL_STATE = 'PROPOSAL_STATE',
  PROPOSAL_AND_LEARNING_STATE = 'PROPOSAL_AND_LEARNING_STATE',
}