// src/student/model/vo/LearningState.ts
export enum LearningState {
  //
  Default = '전체',
  Progress = 'Progress',
  Passed = 'Passed',
  Missed = 'Missed',
  NoShow = 'NoShow',
  Empty = '',

  // Waiting = 'Waiting',
  // Failed = 'Failed',
  // TestPassed = 'TestPassed',
  // TestWaiting = 'TestWaiting',
  // HomeworkWaiting = 'HomeworkWaiting',
}